[Decomp][BattleGround] Cohesion split: extract object spawning to BattleGroundObjects.cpp
